+++ realgui/3rd/box2d/src/dynamics/b2_body.cpp
@@ -546,18 +546,19 @@ void b2Body::Dump()
     b2Dump("{\n");
     b2Dump("  b2BodyDef bd;\n");
     b2Dump("  bd.type = b2BodyType(%d);\n", m_type);
-    b2Dump("  bd.position.Set(%.9g, %.9g);\n", m_xf.p.x, m_xf.p.y);
-    b2Dump("  bd.angle = %.9g;\n", m_sweep.a);
-    b2Dump("  bd.linearVelocity.Set(%.9g, %.9g);\n", m_linearVelocity.x, m_linearVelocity.y);
-    b2Dump("  bd.angularVelocity = %.9g;\n", m_angularVelocity);
-    b2Dump("  bd.linearDamping = %.9g;\n", m_linearDamping);
-    b2Dump("  bd.angularDamping = %.9g;\n", m_angularDamping);
+    b2Dump("  bd.position.Set(%.9g, %.9g);\n", (double)m_xf.p.x, (double)m_xf.p.y);
+    b2Dump("  bd.angle = %.9g;\n", (double)m_sweep.a);
+    b2Dump("  bd.linearVelocity.Set(%.9g, %.9g);\n", (double)m_linearVelocity.x,
+           (double)m_linearVelocity.y);
+    b2Dump("  bd.angularVelocity = %.9g;\n", (double)m_angularVelocity);
+    b2Dump("  bd.linearDamping = %.9g;\n", (double)m_linearDamping);
+    b2Dump("  bd.angularDamping = %.9g;\n", (double)m_angularDamping);
     b2Dump("  bd.allowSleep = bool(%d);\n", m_flags & e_autoSleepFlag);
     b2Dump("  bd.awake = bool(%d);\n", m_flags & e_awakeFlag);
     b2Dump("  bd.fixedRotation = bool(%d);\n", m_flags & e_fixedRotationFlag);
     b2Dump("  bd.bullet = bool(%d);\n", m_flags & e_bulletFlag);
     b2Dump("  bd.enabled = bool(%d);\n", m_flags & e_enabledFlag);
-    b2Dump("  bd.gravityScale = %.9g;\n", m_gravityScale);
+    b2Dump("  bd.gravityScale = %.9g;\n", (double)m_gravityScale);
     b2Dump("  bodies[%d] = m_world->CreateBody(&bd);\n", m_islandIndex);
     b2Dump("\n");
     for (b2Fixture *f = m_fixtureList; f; f = f->m_next)
