+++ realgui/engine/font_mem_matrix.c
@@ -299,7 +299,7 @@ static void rtk_draw_unicode_matrix(mem_char_t *chr, gui_color_t color, uint8_t 
         {
             uint32_t *writebuf = (uint32_t *)dc->frame_buf;
             uint32_t color_back;
-            gui_log("font scale %f , x %d %d , y %d %d \n", scale, x_start, x_end, y_start, y_end);
+            gui_log("font scale %f , x %d %d , y %d %d \n", (double)scale, x_start, x_end, y_start, y_end);
             gui_log(" \n");
             for (uint32_t i = y_start; i < y_end; i++)
             {
